fix(aipr): exit if user declines to push branch
Previously, the script would continue to run even if the user answered 'no' to the prompt to push their branch. This would cause the script to fail later when trying to create a pull request from a branch that does not exist on the remote.

This change ensures the script exits immediately if the push is not performed, preventing subsequent errors.

Signed-off-by: tenfyzhong <tenfy@tenfy.cn>